fp47map.c: revamped for slim buckets, part 1
find, prefetch, new+free, and next are now up to date.
insert needs more work.

The "find32" scheme is now the default (and only).  It seems that
there is no point in updating 32-bit positions, because one first needs
to compare the keys for exact equality, so it has to be a reference.
(And 32-bit references typically cannot be realloc'd or something,
so it has to be a stable reference.)

All the complications with checking the stash are gone, there are
two reasons for this: 1) stashed elements are unlikely to match;
2) to relieve register pressure: after the stash is checked, there is
no need to keep i1 and i2, so those are reused for / transformed into
b1 and b2.

This incorporates some improvements from ../fp47set and ../fp79map,
too hard to track them individually.  The change to i1 = fp >> 32
is not accidental, in increases ILP.