fp47m-sse4.c: initial revision, implemented fp47m_find2_sse4
This is a more elaborate version of "find32 with SSE LeftPack"
(5f73284e33d0).  I've made some benchmarks, the SSE version looks
promising.  find2 isn't much faster than the plain C version, but
it is measurably faster (around 3 out 40+ cycles with bad-proc.c,
logsize=22).  find4 can probably benefit more, we'll see about that.

I've also been studying whether AVX2 can further improve performance.
It turns out that the AVX2 version of LeftPack is very slow on Ryzen.
https://stackoverflow.com/questions/36932240/avx2-what-is-the-most-efficient-way-to-pack-left-based-on-a-mask

I've also been evaluating an AVX version with 128-bit registers where
pshufb is replaced with pmovzxbd+vpermilps.  The benefit is that the
look-up table becomes smaller, 4*16 instead of 16*16 bytes.  There
is otherwise no measurable difference.  But there is also a downside:
with pshufb, we can do "clean packing" with which unused high lanes
are zeroed out.  This is impossible with vpermilps, it sets every output
lane to some of the input lanes.  For find2, it doesn't matter whether
mpos[n] is zero.  But clean packing is indispensable if we want to
resize the structure without branching (we need to load one bucket and
store two buckets, the code isn't written yet, but it seems doable).

Also, the code doesn't seem to benefit from non-destructive VEX encoding.
It may save just a few movdqa instructions, but moves are cheap.  Therefore,
I settle on SSE4 as the main (and probably the only) version for x86.

This also factors fp47m-common.h and fixes a bug in "#define likely".