fp47m-sse4.c: check the mask before leftpack in find4
I double-checked that the benefit is measurable: user time of bad-proc.c
goes down.  (Benchmarking is hard, and microbenchmarks are often flawed,
but user time cannot be feigned.)  find4 benefits more than find2,
probably due to overlapping stores.  We store xpos2 atop of xpos1.
The second store has a penalty associated with it, because the CPU has
to reassemble it with the first store.  The penalty can mean that the
subsequent load from mpos[0], immediately after ret, has to wait a dozen
or so cycles.

Update: applies only to find4.  With find2, the difference is barely
measurable, and probably not statistically significant.  I measure on
Haswell, where the branch misprediction penalty is 15c.  It is 17c on
Skylake, 19c on Ryzen.  So branches should not be added lightly.