Revert "fp47m-sse4.c: check the mask before leftpack in find4"
I was wrong.  The delay of 40+ cycles is too big to be explained by
overlapping stores.  It turns out that the problem is due to malloc.
For an mmap'd chunk, malloc/realloc returns (page + 16).  With 16+16
byte buckets, half the buckets cross the cache line boundary.  In other
words, loading xpos after xtag can cause a delay even despite prefetching.