fp47m-sse4.c: insert[24] need to check the fill factor
With high fill factor, insertions take much longer.  Preliminary
analysis shows that it is better to limit the fill factor directly
rather than reduce maxkick.  The limit should be 75% for insert2
and 87.5% for insert4.  In other words, 1 slot per two buckets should
be free on average, so that the probability of a successful insertion
into any bucket is roughly >= 1/2, and  the initial insertion (before
kicking) succeeds with probability >= 3/4.

I'll add the actual check later, after the rest of the code is written
and debugged.