test-fp47map.c: initial revision
We need to be absolutely sure that the hash map is able to find the
items that we've inserted into it.  Currently this is not the case:
it works with sse4, but fails with the scalar version.

Note that one of the design principles is that internally fp47map
is completely deterministic.  Thus we should be able to hash the
internal state and get the same value with both sse4 and scalar code.
This is contrary to what other implementations try to do.  E.g. in
Abseil, they use ASLR / mix in a stack address.  Instead, I follow
the principle "all the security lies in the key, not in the algorithm".
This means that the hash function should be seeded with AT_RANDOM, and
the algorithm should do only one thing.