new maxkick and full2/full4 stuff
With bigger tables, the fill factor drops steadily (below 77% with
bsize=2).  So instead of maxkick=2*logsize, I now try logsize*log2(logsize).
The average fill factor now settles down at about 84.1% with bsize=2 and
94.8% with bsize=4 (before the first failure; resp. 84.6% and 95.1% before
the second failure).

With high fill factor, insertions take much longer.  Therefore I
explicitly limit the fill factor to (1+9/16)/2=78.125% with bsize=2
and (3+5/8)/4=90.625% with bsize=4.  This is a bit higher than my
previous theory that the fill factor should be limited to 1/2 free slots
per bucket (75% and 87.5%).  With bsize=4, the limit is still higher.
This is because with bsize=2, there are more loops in the graph which
are harder to break (because of limited eviction choices).  In other
words, insert4 will fail more often than insert2, but kickloop4 will
complete sooner than kickloop2.