make insert4 fail if fill factor < 50%
The failure means that the table is left in an inconsistent state (with
an unrelated entry kicked out), and the only option left to the caller
is to rebuild the table from scratch, rehashing the data with a
different seed.  The probability of such a failure is very small though,
and for many purposes negligible.  So the caller may as well abort.

We shouldn't report insert2 failures in the same way, because the
probability of cycles in its cuckoo graph is much higher.  So insert2
should resize 2->4 entries per bucket unconditionally.